feat: allow dictionary substitutions anywhere in test definition
The "body" item can now specify JSON with substititon values; these
are applied when the JSON file is first read from disk before decoding
the JSON.

This eliminates the need to use the "file" definition for a payload to
support dictionary substitutions. A file is still useful if it was
generated during a test, or is a common payload being shared by
many tests.